Changed var to let. Moved code from vs dir to sql. Removed support for showing file location after file save. (Will be moved to another PR). #746
