+++ src/sql/workbench/common/workspaceActions.ts
@@ -1,0 +1,25 @@
+import { TPromise } from 'vs/base/common/winjs.base';
+import { Action } from 'vs/base/common/actions';
+import { IWindowsService } from 'vs/platform/windows/common/windows';
+
+export class ShowFileInFolderAction extends Action {
+
+	constructor(private path: string, label: string, private windowsService: IWindowsService) {
+		super('showItemInFolder.action.id', label);
+	}
+
+	run(): TPromise<void> {
+		return this.windowsService.showItemInFolder(this.path);
+	}
+}
+
+export class OpenFileInFolderAction extends Action {
+
+	constructor(private path: string, label: string, private windowsService: IWindowsService) {
+		super('showItemInFolder.action.id', label);
+	}
+
+	run() {
+		return this.windowsService.openExternal(this.path);
+	}
+}
